feat:converting

+++ benches/signing.rs
@@ -46,7 +46,7 @@ pub fn build_submit_order_dto_for_bench(
     time_in_force: TimeInForce,
     post_only: bool,
     reduce_only: bool,
-    expires_at: Option<f64>,
+    expires_at: Option<i64>,
 ) -> Result<SubmitOrderDto, Box<dyn std::error::Error>> {
     if !client.product_hashmap.contains_key(ticker) {
         return Err(format!("Ticker {ticker} not found").into());
